requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.0
begins==0.9
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
eyeD3==0.8.5
future==0.16.0
greenlet==0.4.13
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
lxml==4.2.1
msgpack==0.5.6
neovim==0.2.6
numpy==1.14.3
pandas==0.22.0
parsedatetime==2.4
parso==0.2.1
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.2
python-magic==0.4.15
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tqdm==4.23.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xlrd==1.1.0
